hash changes

+++ dataStructure/hash.h
@@ -4,17 +4,26 @@
 
 #ifndef PROJECT_3_HASH_H
 #define PROJECT_3_HASH_H
+
 #include "../utils/name.h"
 #include <stdbool.h>
-typedef struct hashtable{
+
+typedef struct hashtable {
     Name *names;
     int currentSize;
     int capacity;
-}hash;
-long long int hashPos(Name);
-hash createTable(int );
-void insertTable(hash *, Name );
-bool searchHash(hash *, Name );
-void deleteHash(hash *,Name );
+} hash;
+
+int hashPos(Name);
+
+hash createTable(int);
+
+void insertTable(hash *, Name);
+
+bool searchHash(hash *, Name);
+
+void deleteHash(hash *, Name);
+
 void listTable(hash *);
+
 #endif //PROJECT_3_HASH_H
